--- a/src/System.CommandLine.Tests/SymbolSetExtensions.cs
+++ b/src/System.CommandLine.Tests/SymbolSetExtensions.cs
@@ -1,17 +1,17 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.CommandLine.Collections;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace System.CommandLine.Tests
 {
     internal static class SymbolSetExtensions
     {
-        internal static Symbol GetByAlias(this SymbolSet symbolSet, string alias)
+        internal static Symbol GetByAlias(this IEnumerable<Symbol> symbolSet, string alias)
             => symbolSet.SingleOrDefault(symbol => symbol.Name.Equals(alias) || symbol is IdentifierSymbol id && id.HasAlias(alias));
 
-        internal static bool ContainsAlias(this SymbolSet symbolSet, string alias)
+        internal static bool ContainsAlias(this IEnumerable<Symbol> symbolSet, string alias)
             => symbolSet.OfType<IdentifierSymbol>().Any(symbol => symbol.HasAlias(alias));
     }
 }