--- a/src/System.CommandLine/Symbol.cs
+++ b/src/System.CommandLine/Symbol.cs
@@ -2,10 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
-using System.CommandLine.Collections;
-using System.CommandLine.Parsing;
 using System.CommandLine.Completions;
-using System.Linq;
 
 namespace System.CommandLine
 {