… src/System.CommandLine.ApiCompatibility.Tests/ApiCompatibilityApprovalTests.System_CommandLine_api_is_not_changed.approved.txt
… src/System.CommandLine.Tests/SymbolSetTests.cs
… src/System.CommandLine/Collections/SymbolSet.cs
… src/System.CommandLine/Parsing/ParseResultVisitor.cs
… src/System.CommandLine/Parsing/StringExtensions.cs
… src/System.CommandLine/SymbolExtensions.cs
… src/System.CommandLine/SymbolSetExtensions.cs