@jonsequitur this time the perf gains are smaller (mostly allocations), I think I've reached the point where the only part where I can find some potential perf bottlenecks is argument conversion (not included in this PR, I've created another branch with large experiment).


|        Method |      Args | Mean before | Mean after | Allocated before | Allocated  after|
|-------------- |---------- |------------:|-----------:|-----------------:|----------------:|
|  DefaultsSync | String[0] |   27.654 us |  27.682 us |         11,850 B |        11,521 B |
| DefaultsAsync | String[0] |   27.608 us |  27.775 us |         11,849 B |        11,521 B |
|   MinimalSync | String[0] |    2.646 us |   2.592 us |          4,961 B |         4,665 B |
|  MinimalAsync | String[0] |    2.766 us |   2.719 us |          4,961 B |         4,664 B |
|  DefaultsSync | String[4] |   31.348 us |  30.990 us |         15,152 B |        14,624 B |
| DefaultsAsync | String[4] |   31.433 us |  31.284 us |         15,152 B |        14,624 B |
|   MinimalSync | String[4] |    4.805 us |   4.606 us |          8,264 B |         7,768 B |
|  MinimalAsync | String[4] |    4.981 us |   4.628 us |          8,264 B |         7,768 B |