@jonsequitur 2021-12-23
re: argument conversion, there's some reflection code left there that we'd ideally like to move out of the core library to remove trimming warnings. Right now, any class with a constructor taking a single string (found via reflection) can be bound. We can move this open-ended implementation to `NamingConventionBinder` and keep support for a few key types if needed by adding converters for them.