@adamsitnik 2022-01-26
Closing in favor of https://github.com/dotnet/command-line-api/pull/1601